// src/shared/Shared.d.ts
export declare type Variant = "primary" | "secondary" | "success" | "danger" | "warning" | "info" | "dark" | "light";
export declare type ButtonVariant = Variant | "link" | "outline-primary" | "outline-secondary" | "outline-success" | "outline-danger" | "outline-warning" | "outline-info" | "outline-dark" | "outline-light";